+++ src/replica/construct.rs
@@ -280,6 +280,7 @@ impl<K: Key + Hash, V: Value> Replica<K, V> {
                 max_peers: PeerCap::new(config.max_peers),
                 coalesce_window: Arc::new(RwLock::new(config.coalesce_window)),
                 coalesce_pending: Arc::new(RwLock::new(HashMap::new())),
+                changes_since_snapshot: Arc::new(AtomicUsize::new(0)),
             }),
         }
     }